Full Stack app created

+++ routes/ideas.js
@@ -2,29 +2,7 @@ const express = require("express");
 const router = express.Router();
 const Idea = require("../models/Idea");
 
-const ideas = [
-  {
-    id: 1,
-    text: "Positive NewsLetter, a newsletter that only shares positive, uplifting news",
-    tag: "Technology",
-    username: "TonyStark",
-    date: "2022-01-02",
-  },
-  {
-    id: 2,
-    text: "Milk cartons that turn a different color the older that your milk is getting",
-    tag: "Inventions",
-    username: "SteveRogers",
-    date: "2022-01-02",
-  },
-  {
-    id: 3,
-    text: "ATM location app which lets you know where the closest ATM is and if it is in service",
-    tag: "Software",
-    username: "BruceBanner",
-    date: "2022-01-02",
-  },
-];
+const ideas = [];
 
 //Get all ideas
 router.get("/", async (req, res) => {
@@ -44,7 +22,7 @@ router.get("/:id", async (req, res) => {
     res.json({ success: true, data: idea });
   } catch (error) {
     console.log(error);
-    res.status(500).json({ success: true, error: "Something went wrong" });
+    res.status(500).json({ success: false, error: "Something went wrong" });
   }
 });
 
@@ -70,31 +48,53 @@ router.post("/", async (req, res) => {
 
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    //Match the usernames
+    const idea = await Idea.findById(req.params.id);
+
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    //username don't match
+    res.status(403).json({
+      success: false,
+      error: "You are not authorize to delete this resource",
+    });
   } catch (error) {
     console.log(error);
-    res.status(500).json({ success: true, error: "Something went wrong" });
+    res.status(500).json({ success: false, error: "Something went wrong" });
   }
 });
 
 //Delete an idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    //Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    //username don't match
+    res.status(403).json({
+      success: false,
+      error: "You are not authorize to delete this resource",
+    });
   } catch (error) {
     console.log(error);
-    res.status(500).json({ success: true, error: "Something went wrong" });
+    res.status(500).json({ success: false, error: "Something went wrong" });
   }
 });
